Kan jo like gjerne trekkje ut omgjøring etter klage i eiga service også når vi først er i gang
