+++ apps/etterlatte-behandling/src/main/kotlin/behandling/revurdering/RevurderingService.kt
@@ -11,14 +11,12 @@ import no.nav.etterlatte.behandling.domain.Revurdering
 import no.nav.etterlatte.behandling.domain.toBehandlingOpprettet
 import no.nav.etterlatte.behandling.domain.toStatistikkBehandling
 import no.nav.etterlatte.behandling.hendelse.HendelseDao
-import no.nav.etterlatte.behandling.klage.KlageService
 import no.nav.etterlatte.behandling.kommerbarnettilgode.KommerBarnetTilGodeService
 import no.nav.etterlatte.libs.common.Vedtaksloesning
 import no.nav.etterlatte.libs.common.behandling.BehandlingHendelseType
 import no.nav.etterlatte.libs.common.behandling.BehandlingStatus
 import no.nav.etterlatte.libs.common.behandling.BehandlingType
 import no.nav.etterlatte.libs.common.behandling.BoddEllerArbeidetUtlandet
-import no.nav.etterlatte.libs.common.behandling.Klage
 import no.nav.etterlatte.libs.common.behandling.Persongalleri
 import no.nav.etterlatte.libs.common.behandling.Prosesstype
 import no.nav.etterlatte.libs.common.behandling.RevurderingInfo
@@ -26,14 +24,11 @@ import no.nav.etterlatte.libs.common.behandling.Revurderingaarsak
 import no.nav.etterlatte.libs.common.behandling.Utlandstilknytning
 import no.nav.etterlatte.libs.common.behandling.Virkningstidspunkt
 import no.nav.etterlatte.libs.common.feilhaandtering.IkkeTillattException
-import no.nav.etterlatte.libs.common.feilhaandtering.InternfeilException
 import no.nav.etterlatte.libs.common.feilhaandtering.UgyldigForespoerselException
 import no.nav.etterlatte.libs.common.grunnlag.Grunnlagsopplysning
-import no.nav.etterlatte.libs.common.oppgave.OppgaveIntern
 import no.nav.etterlatte.libs.common.oppgave.OppgaveKilde
 import no.nav.etterlatte.libs.common.oppgave.OppgaveType
 import no.nav.etterlatte.libs.common.tidspunkt.Tidspunkt
-import no.nav.etterlatte.libs.ktor.token.Saksbehandler
 import no.nav.etterlatte.oppgave.OppgaveService
 import org.slf4j.LoggerFactory
 import java.time.LocalDateTime
@@ -87,7 +82,6 @@ class RevurderingService(
     private val hendelseDao: HendelseDao,
     private val kommerBarnetTilGodeService: KommerBarnetTilGodeService,
     private val revurderingDao: RevurderingDao,
-    private val klageService: KlageService,
     private val aktivitetspliktDao: AktivitetspliktDao,
     private val aktivitetspliktKopierService: AktivitetspliktKopierService,
 ) {
@@ -253,74 +247,6 @@ class RevurderingService(
         val revurderingInfo = RevurderingInfo.RevurderingAarsakAnnen(fritekstAarsak)
         lagreRevurderingInfo(behandlingId, RevurderingInfoMedBegrunnelse(revurderingInfo, null), saksbehandlerIdent)
     }
-
-    fun opprettOmgjoeringKlage(
-        sakId: Long,
-        oppgaveIdOmgjoering: UUID,
-        saksbehandler: Saksbehandler,
-    ): Revurdering {
-        val omgjoeringsoppgave = oppgaveService.hentOppgave(oppgaveIdOmgjoering)
-        if (omgjoeringsoppgave.type != OppgaveType.OMGJOERING) {
-            throw FeilIOmgjoering.IngenOmgjoeringsoppgave()
-        }
-        if (omgjoeringsoppgave.status.erAvsluttet()) {
-            throw FeilIOmgjoering.OmgjoeringsOppgaveLukket(omgjoeringsoppgave)
-        }
-        if (omgjoeringsoppgave.sakId != sakId) {
-            throw FeilIOmgjoering.OppgaveOgSakErForskjellig(sakId, omgjoeringsoppgave)
-        }
-        if (omgjoeringsoppgave.saksbehandler?.ident != saksbehandler.ident) {
-            throw FeilIOmgjoering.SaksbehandlerHarIkkeOppgaven(saksbehandler, omgjoeringsoppgave)
-        }
-
-        val klageId = UUID.fromString(omgjoeringsoppgave.referanse)
-        val klagenViOmgjoerPaaGrunnAv =
-            klageService.hentKlage(klageId)
-                ?: throw InternfeilException(
-                    "Omgjøringsoppgaven med id=${omgjoeringsoppgave.id} peker på en " +
-                        "klageId=${omgjoeringsoppgave.referanse} som vi ikke finner. " +
-                        "Her har noe blitt koblet feil, og må ryddes opp i.",
-                )
-
-        val behandlingSomOmgjoeresId =
-            klagenViOmgjoerPaaGrunnAv.formkrav
-                ?.formkrav
-                ?.vedtaketKlagenGjelder
-                ?.behandlingId
-                ?.let { UUID.fromString(it) }
-                ?: throw FeilIOmgjoering.ManglerBehandlingForOmgjoering(klagenViOmgjoerPaaGrunnAv)
-        val behandlingSomOmgjoeres =
-            behandlingDao.hentBehandling(behandlingSomOmgjoeresId)
-                ?: throw FeilIOmgjoering.ManglerBehandlingForOmgjoering(klagenViOmgjoerPaaGrunnAv)
-
-        val persongalleri = runBlocking { grunnlagService.hentPersongalleri(behandlingSomOmgjoeres.id) }
-        return opprettRevurdering(
-            sakId = sakId,
-            persongalleri = persongalleri,
-            forrigeBehandling = behandlingSomOmgjoeresId,
-            mottattDato =
-                klagenViOmgjoerPaaGrunnAv.innkommendeDokument
-                    ?.mottattDato
-                    ?.atStartOfDay()
-                    ?.toString(),
-            prosessType = Prosesstype.MANUELL,
-            kilde = Vedtaksloesning.GJENNY,
-            revurderingAarsak = Revurderingaarsak.OMGJOERING_ETTER_KLAGE,
-            virkningstidspunkt = behandlingSomOmgjoeres.virkningstidspunkt,
-            utlandstilknytning = behandlingSomOmgjoeres.utlandstilknytning,
-            boddEllerArbeidetUtlandet = behandlingSomOmgjoeres.boddEllerArbeidetUtlandet,
-            begrunnelse = "Omgjøring på grunn av klage",
-            fritekstAarsak = omgjoeringsoppgave.merknad,
-            saksbehandlerIdent = saksbehandler.ident,
-            relatertBehandlingId = klagenViOmgjoerPaaGrunnAv.id.toString(),
-        ).oppdater().also {
-            oppgaveService.ferdigStillOppgaveUnderBehandling(
-                referanse = klagenViOmgjoerPaaGrunnAv.id.toString(),
-                type = OppgaveType.OMGJOERING,
-                saksbehandler = saksbehandler,
-            )
-        }
-    }
 }
 
 data class RevurderingOgOppfoelging(
@@ -339,40 +265,4 @@ data class RevurderingOgOppfoelging(
     fun behandlingId() = revurdering.id
 
     fun sakType() = revurdering.sak.sakType
-}
-
-sealed class FeilIOmgjoering {
-    class IngenOmgjoeringsoppgave :
-        UgyldigForespoerselException("INGEN_OMGJOERINGSOPPGAVE", "Mottok ikke en gyldig omgjøringsoppgave")
-
-    class OmgjoeringsOppgaveLukket(
-        oppgave: OppgaveIntern,
-    ) : UgyldigForespoerselException(
-            "OMGJOERINGSOPPGAVE_LUKKET",
-            "Oppgaven ${oppgave.id} har status ${oppgave.status}.",
-        )
-
-    class OppgaveOgSakErForskjellig(
-        sakId: Long,
-        oppgave: OppgaveIntern,
-    ) : UgyldigForespoerselException(
-            "SAK_I_OPPGAVE_MATCHER_IKKE",
-            "Saken det skal omgjøres i har id=$sakId, men omgjøringsoppgaven er i sak med id=${oppgave.sakId}",
-        )
-
-    class SaksbehandlerHarIkkeOppgaven(
-        saksbehandler: Saksbehandler,
-        omgjoeringsoppgave: OppgaveIntern,
-    ) : UgyldigForespoerselException(
-            "SAKSBEHANDLER_HAR_IKKE_OPPGAVEN",
-            "Saksbehandler med ident=${saksbehandler.ident} er ikke saksbehandler i oppgaven med " +
-                "id=$omgjoeringsoppgave (saksbehandler i oppgaven er ${omgjoeringsoppgave.saksbehandler?.ident}).",
-        )
-
-    class ManglerBehandlingForOmgjoering(
-        klage: Klage,
-    ) : InternfeilException(
-            "Klagen med id=${klage.id} har laget en omgjøringsoppgave men vi finner ikke behandlingen som skal omgjøres." +
-                " Noe galt har skjedd i ferdigstillingen av denne klagen, eller dette er ikke et behandlingsvedtak.",
-        )
 }
